add a metric name to limit number of returned queries (#1) (#43)
In large environments the topic, bookies an zookeeper dashboards can't load and cause to a browser on the client side to freeze.

It happens because the dashboard tries to retrieve the list of variables such as list of jobs, list of namespaces and so on. It does it by queries to Prometheus,
As mentioned before in large environments these queries sometimes can return up to 200k rows for each variable.

By limiting these queries by specifying a metric name we can significantly improve the response time and prevent from the dashboard to be stuck.