Auto Recovery Dashboard Should Use Sum Instead of Count for Total Bytes (#58)
The recovery dashboard has a panel titled `Replication read throughput (bytes/s)` and `Replication write throughput (bytes/s)`. Each of these use the `count` metric instead of the `sum` metric. Given that we want to know how many bytes have been read or written, we want the total bytes, not the number of times the metric has been incremented.

Further, I updated the units to bytes on the y-axis of each panel, and I set the minimum to 0 because it is a rate on an increasing counter, so it'll always be non-negative.